refactor(api): extract ApiClient to shared lib and update imports
